:sparkles: feat: Realizar transacciones (Depositos, retiros, pagos y cargos) y obtener movimientos por porducto

+++ src/main/java/com/bm_nttdata/transaction_ms/entity/Transaction.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -23,8 +24,8 @@ public class Transaction {
     private String id;
     private TransactionTypeEnum transactionType;
     private ProductTypeEnum productType;
-    private Double amount;
-    private Double commission;
+    private BigDecimal amount;
+    private BigDecimal commission;
     private String sourceAccountId;
     private String targetAccountId;
     private String creditId;
